+++ cryptography/hedera-cryptography-blskeygen/src/main/java/com/hedera/cryptography/blskeygen/CliCommand.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.cryptography.blskeygen;
+
+/**
+ * The KeyGen commands
+ */
+public enum CliCommand {
+    /**
+     * Print the help message
+     */
+    PRINT_HELP,
+    /**
+     * Generate a key pair
+     */
+    GENERATE_KEYS,
+    /**
+     * Generate a public key from a private key
+     */
+    GENERATE_PUBLIC_KEY
+}
