+++ cryptography/hedera-cryptography-blskeygen/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+/**
+ * This module provides a cryptography utility to create EC PublicKeys and EC PrivateKeys.
+ */
+module com.hedera.cryptography.blskeygen {
+    requires com.hedera.cryptography.pairings.api;
+    requires com.hedera.cryptography.pairings.signatures;
+    requires com.google.protobuf;
+    requires static transitive com.github.spotbugs.annotations;
+}
